requirements.txt
matplotlib==3.7.1
natsort==8.3.1
numpy==1.24.3
pandas==2.0.1
pydicom==2.3.1
PyYAML==6.0
Requests==2.31.0
scikit_image==0.19.3
scikit_learn==1.2.2
